Migrate into big integer

+++ database/migrations/2015_09_15_064722_create_educations_table.php
@@ -13,13 +13,13 @@ class CreateEducationsTable extends Migration
     public function up()
     {
         Schema::create('educations', function (Blueprint $table) {
-            $table->increments('id');
+            $table->bigIncrements('id');
             $table->string('school');
             $table->integer('degree');
             $table->string('field_of_study')->nullable();
             $table->integer('start_year');
             $table->integer('end_year')->nullable();
-            $table->integer('volunteer_id')->index()->unsigned();    // foregin key
+            $table->bigInteger('volunteer_id')->index()->unsigned();    // foregin key
             $table->dateTime('created_at');
             $table->dateTime('updated_at');
